pyproject.toml
[tool.poetry]
name = "exchange rate bot"
version = "0.1.0"
description = ""
authors = ["vshadrin1991 <v.shadrin@setplex.com>"]
readme = "README.md"
packages = [{include = "exchange"}]

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
pytelegrambotapi = "^4.12.0"
coloredlogs = "^15.0.1"
sqlalchemy = "^2.0.18"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.3"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"

[tool.black]
skip-string-normalization = true

[tool.isort]
profile = "black"
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
